requirements.txt
langchain>=0.0.200
langchain-openai>=0.0.2
openai>=1.0.0
requests>=2.31.0
schedule>=1.2.0
python-dotenv>=1.0.0
pytz>=2023.3